refactoring the level processor
