+++ src/assets/levels/processed/level-0/images.js
@@ -1,14 +1,20 @@
-import BackgroundGradient from './background-with-planets.png';
+
+  
+
+  import BackgroundGradient from './background-with-planets.png';
 import BackgroundSea from './mountain_and_sea.png';
 import BackgroundClouds from './clouds.png';
+  
 
-import rockMossPlantsDoors from './rock-moss-plants-doors.png';
+  import rockMossPlantsDoors from './rock-moss-plants-doors.png';
 import inputTiles from './input-tiles.png';
-export default {
-  BackgroundGradient: BackgroundGradient,
-  BackgroundSea: BackgroundSea,
-  BackgroundClouds: BackgroundClouds,
-
-  'rock-moss-plants-doors': rockMossPlantsDoors,
-  'input-tiles': inputTiles
-};
+  export default {
+    
+'BackgroundGradient':BackgroundGradient,
+'BackgroundSea':BackgroundSea,
+'BackgroundClouds':BackgroundClouds,
+    
+'rock-moss-plants-doors':rockMossPlantsDoors,
+'input-tiles':inputTiles
+  };
+
